- Logica inicial para trigger de token alpha

+++ Transactions/SyncronizationBot.Domain/Model/Database/TokenAlphaWallet.cs
@@ -8,7 +8,7 @@ namespace SyncronizationBot.Domain.Model.Database
         public Guid? TokenAlphaId { get; set; }
         public Guid? WalletId { get; set; }
         public int? NumberOfBuys { get; set; }
-        public int? ValueSpentSol { get; set; }
+        public decimal? ValueSpentSol { get; set; }
         public decimal? ValueSpentUSDC { get; set; }
         public decimal? ValueSpentUSDT { get; set; }
         public virtual TokenAlpha? TokenAlpha { get; set; }
